requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.1
django-database-url==1.0.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
sympy==1.12
numpy==1.26.2
scipy==1.11.4
djangorestframework==3.14.0
django-cors-headers==4.3.1